rename certs -> customerCerts
